requirements.txt
Django==2.2.7
djangorestframework==3.10.3 
django-cors-headers==3.1.1
Pygments==2.4.2  
tensorflow==1.5.0
keras==2.2.4
opencv-python==4.1.1.26
pillow==6.2.1
